refactor submodule names
