+++ tipg/dependencies.py
@@ -10,7 +10,7 @@ from pygeofilter.ast import AstType
 from pygeofilter.parsers.cql2_json import parse as cql2_json_parser
 from pygeofilter.parsers.cql2_text import parse as cql2_text_parser
 
-from tipg.dbmodel import Collection, Database
+from tipg.collections import Catalog, Collection
 from tipg.errors import InvalidBBox, MissingCollectionCatalog, MissingFunctionParameter
 from tipg.resources import enums
 from tipg.settings import TMSSettings
@@ -44,9 +44,7 @@ def CollectionParams(
     assert collection_pattern.groupdict()["schema"]
     assert collection_pattern.groupdict()["collection"]
 
-    collection_catalog: Database = getattr(
-        request.app.state, "collection_catalog", None
-    )
+    collection_catalog: Catalog = getattr(request.app.state, "collection_catalog", None)
     if not collection_catalog:
         raise MissingCollectionCatalog("Could not find collections catalog.")
